+++ project-scoto/Assets/src/rodney/Code/WeaponManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c6d3a15faff9f2e47b69f8e366f469d8
+guid: 90d304cb104a85b40a2d960b3ab5a50d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
